move solution classes to client package #1231
