+++ commercetools-models/src/test/resources/META-INF/services/io.sphere.sdk.client.SolutionInfoSupplier
